Change parser to read more data from the network when necessary
Before this commit the parser was reading from the network once and
hoping to catch a full message. However, that is a flaw and in a real world scenario
the network will not always reply with all of the data for a single recv() call.

This commit rewrites the parser to call recv() on-demand, based on the Redis protocol message