Create new sorted set score structure (binary)
The new format has an 8-byte header (count) followed by the elements data.
Each element is represented by a 4-byte header and a data section.
The 4-byte header contains the size of the data section.
The data section is the raw data (no serialization).

Example of a score file with 2 elements ("hello" and "world"):

    0x0002   0x05 hello     0x05 world
     count   size data      size data
      2       5   "hello"   5    "world"