Fix TCP reads to ignore half-baked messages
A chunk is read from the network and added to a buffer (recv()),
then if the buffer doesn't have enough information, the handler
waits for the next `handle_read` call to get more information into the buffer.
This works with simultaneous clients.

The Redis codebase was the guide for this commit.