Refactor magic ziplist magic numbers to constants
Hashes with 512 or fewer entries will use a ziplist encoding by default, which changes
the behavior of HSCAN to return all entries, regardless of the COUNT parameter.
The same applies to sorted sets with 128 or fewer entries.

The `+ 100` is to avoid issues with `range()` & off-by-one errors (simpler than adding `+ 1` to the `range()` calls).