Use `None` to flag missing arguments instead of the empty string
The empty string was a shortcut because it is a falsy value in Python, but it did not
express the right intentions and required a hack to work with plyvel (it does not
accept key_prefix and start arguments together).

This commit makes the intentions clear but requires some `None` checks.