Send remaining `send` bytes in the future
Sometimes `socket.send` can't send all the data at once. Before this commit,
the client would miss data because there was no consequent attempts
to send what was missing. We saw issues with `HSCAN` and 1MB replies.

Using `socket.sendall` is not a good alternative because the client
may not be ready to accept all the data at once (it will raise an exception).
It is better to let the client request for more data (`handle_write`) instead
of calling `socket.sendall`.

This commit implements an event-driven solution with a buffer,
similar to the Redis implementation.
This work is based on `asyncore.dispatcher_with_send` but uses `bytearray` for
better performance (similar to the `dredis/parser.py` code).