Optimize array parsing to avoid re-parsing
There was too much re-parsing when dealing with large inputs that required multiple reads from the network.
@jimjshields found this when sending about 10MB of data to dredis (it was hanging for a long time). That was hanging dredis because it was re-parsing the partial input every time there was new data.
For example, if a client sent the following sequence of commands to parse:
1. "*1\r\n"
2. "$4\r\n"
3. "PING\r\n"

After receiving payload #1, dredis would append it to a buffer and parse the buffer ("*1\r\n"). It is incomplete, so it moves on.
After receiving payload #2, dredis would append it to a buffer and parse the buffer (now the buffer is "*1\r\n$4\r\n"). It is incomplete, so it moves on.
After receiving payload #3, dredis would append it to a buffer and parse the buffer (nw the buffer is "*1\r\n$4\r\nPING\r\n"). It is finally complete, so it executes the command PING.
If you follow this approach for a command with 50,000 elements, it would re-parse all elements until it reaches the final 50,000th element. A much better approach would be to save every parsed element and wait until the last element to execute the command, but never re-parse valid data.

This commit changes the parser to never re-parse valid data. Using the previous example with the current implementation, after each new payload, dredis will only that and never re-parse the previous valid payloads.