Change RENAME to take advantage of key IDs
This new behavior will only touch a couple of keys and is
much more efficient than renaming all backend keys.
The new code is O(1) instead of O(N) where N is the number
of backend storage key-value pairs for a given dredis key;
it is finally renaming a pointer.